code review, add `group_id` attribute
