+++ docs/data-sources/freeform_resource_group_generator.md
@@ -1,12 +1,12 @@
 ---
-page_title: "apstra_freeform_group_generator Data Source - terraform-provider-apstra"
+page_title: "apstra_freeform_resource_group_generator Data Source - terraform-provider-apstra"
 subcategory: "Reference Design: Freeform"
 description: |-
   This data source provides details of a specific Freeform Group Generator.
   At least one optional attribute is required.
 ---
 
-# apstra_freeform_group_generator (Data Source)
+# apstra_freeform_resource_group_generator (Data Source)
 
 This data source provides details of a specific Freeform Group Generator.
 
@@ -16,8 +16,8 @@ At least one optional attribute is required.
 ## Example Usage
 
 ```terraform
-# This example creates an group Generator within a
-# preexisting resource group in a Freeform Blueprint.
+# This example creates a Resource Group Generator under a
+# Resource Group in a Freeform Blueprint.
 #
 # After creating the Group Generator, the data source is invoked to look up
 # the details.
@@ -29,30 +29,31 @@ resource "apstra_freeform_resource_group" "fizz_grp" {
 }
 
 # Create a resource generator scoped to target all systems in the blueprint.
-resource "apstra_freeform_group_generator" "test_group_gen" {
-  blueprint_id   = "f1b86583-9139-49ed-8a3c-0490253e006e"
-  name           = "test_res_gen"
-  scope          = "node('system', name='target')"
+resource "apstra_freeform_resource_group_generator" "test_group_gen" {
+  blueprint_id = "f1b86583-9139-49ed-8a3c-0490253e006e"
+  group_id     = apstra_freeform_resource_group.fizz_grp.id
+  name         = "test_res_gen"
+  scope        = "node('system', name='target')"
 }
 
-# Invoke the resource generator data source
-data "apstra_freeform_group_generator" "test_group_gen" {
-  blueprint_id   = "f1b86583-9139-49ed-8a3c-0490253e006e"
-  id = apstra_freeform_group_generator.test_group_gen.id
+# Invoke the resource group generator data source
+data "apstra_freeform_resource_group_generator" "test_group_gen" {
+  blueprint_id = "f1b86583-9139-49ed-8a3c-0490253e006e"
+  id           = apstra_freeform_resource_group_generator.test_group_gen.id
 }
 
 # Output the data source so that it prints on screen
-output "test_group_generator_out" {
-  value = data.apstra_freeform_group_generator.test_group_gen
+output "test_resource_group_generator_out" {
+  value = data.apstra_freeform_resource_group_generator.test_group_gen
 }
 
 # The output looks like this:
-#test_group_generator_out = {
-#  "blueprint_id" = "f1b86583-9139-49ed-8a3c-0490253e006e"
-#  "id" = "wOQS9qZezzRJCgyvxwY"
-#  "name" = "test_res_gen"
-#  "scope" = "node('system', name='target')"
-#}
+# test_resource_group_generator_out = {
+#   "blueprint_id" = "f1b86583-9139-49ed-8a3c-0490253e006e"
+#   "id" = "wOQS9qZezzRJCgyvxwY"
+#   "name" = "test_res_gen"
+#   "scope" = "node('system', name='target')"
+# }
 ```
 
 <!-- schema generated by tfplugindocs -->
@@ -69,5 +70,6 @@ output "test_group_generator_out" {
 
 ### Read-Only
 
+- `group_id` (String) Resource Group the Group Generator belongs to.
 - `scope` (String) Scope is a graph query which selects target nodes for which Groups should be generated.
 Example: `node('system', name='target', label=aeq('*prod*'))`
