--- docs/resources/freeform_group_generator.md
@@ -1,70 +1,0 @@
----
-page_title: "apstra_freeform_group_generator Resource - terraform-provider-apstra"
-subcategory: "Reference Design: Freeform"
-description: |-
-  This resource creates a Group Generator in a Freeform Blueprint.
----
-
-# apstra_freeform_group_generator (Resource)
-
-This resource creates a Group Generator in a Freeform Blueprint.
-
-
-## Example Usage
-
-```terraform
-# This example creates an group Generator within a
-# preexisting resource group in a Freeform Blueprint.
-#
-# After creating the Group Generator, the data source is invoked to look up
-# the details.
-
-# Create a resource group in a preexisting blueprint.
-resource "apstra_freeform_resource_group" "fizz_grp" {
-  blueprint_id = "f1b86583-9139-49ed-8a3c-0490253e006e"
-  name         = "fizz_grp"
-}
-
-# Create a resource generator scoped to target all systems in the blueprint.
-resource "apstra_freeform_group_generator" "test_group_gen" {
-  blueprint_id   = "f1b86583-9139-49ed-8a3c-0490253e006e"
-  name           = "test_res_gen"
-  scope          = "node('system', name='target')"
-}
-
-# Invoke the resource generator data source
-data "apstra_freeform_group_generator" "test_group_gen" {
-  blueprint_id   = "f1b86583-9139-49ed-8a3c-0490253e006e"
-  id = apstra_freeform_group_generator.test_group_gen.id
-}
-
-# Output the data source so that it prints on screen
-output "test_group_generator_out" {
-  value = data.apstra_freeform_group_generator.test_group_gen
-}
-
-# The output looks like this:
-#test_group_generator_out = {
-#  "blueprint_id" = "f1b86583-9139-49ed-8a3c-0490253e006e"
-#  "id" = "wOQS9qZezzRJCgyvxwY"
-#  "name" = "test_res_gen"
-#  "scope" = "node('system', name='target')"
-#}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `blueprint_id` (String) Apstra Blueprint ID.
-- `name` (String) Freeform Group Generator name as shown in the Web UI.
-- `scope` (String) Scope is a graph query which selects target nodes for which Groups should be generated.
-Example: `node('system', name='target', label=aeq('*prod*'))`
-
-### Read-Only
-
-- `id` (String) ID of the Group Generator within the Freeform Blueprint.
-
-
-
